+++ dejafu-tests/Examples/SearchParty.hs
@@ -21,8 +21,8 @@ import Data.Functor (void)
 import Data.Maybe (fromJust, isNothing)
 
 -- test imports
-import Data.List (permutations)
-import Test.DejaFu (Predicate, Result(..), alwaysTrue2)
+import Data.List (sort)
+import Test.DejaFu (Predicate, Result(..), alwaysSameOn)
 import Test.Framework (Test)
 import Test.Framework.Providers.HUnit (hUnitTestToTests)
 import Test.HUnit (test)
@@ -42,11 +42,8 @@ concFilter :: MonadConc m => m [Int]
 concFilter = unsafeRunFind $ [0..5] @! const True
 
 -- | Check that two lists of results are equal, modulo order.
-checkResultLists :: Eq a => Predicate [a]
-checkResultLists = alwaysTrue2 checkLists where
-  checkLists (Right as) (Right bs) =
-    as `elem` permutations bs
-  checkLists a b = a == b
+checkResultLists :: Ord a => Predicate [a]
+checkResultLists = alwaysSameOn (fmap sort)
 
 -------------------------------------------------------------------------------